finish refactor for highlight block. relates to #504.
