+++ src/blocks/highlight/colors.js
