+++ src/blocks/highlight/controls.js
