+++ src/blocks/highlight/inspector.js
